DEV-1014 Don't process event when outcome isn't OK
Refactor _handle_premis_event to have less nested if-else statements.
The loss of indentation improves readability.

Make get_fragment_metadata and generate_vrt_xml "private" functions
to be consistent with the "private" character of _handle_premis_event.

Refactor test_app.py so that it mocks on RabbitClient level and not
on underlying pika.connection level. The latter needs te be fully
tested in test_rabbit_service.py. As a result the mocking of the
config can also be removed. This leads to less code.