DEV-1440 Change Flask with FastAPI
The incoming premis events are processed synchronously. This process can
take up to more than 10 seconds per event, resulting in a timeout from MH.
Switch to FastAPI in order to handle these requests in a asynchronous way.

Correlation has been removed for now as there is a dependecy on Flask.