DEV-1440 Make Uvicorn logging more uniform
Messages logged by Uvicorn are actually logged twice: once by the Uvicorn
logger and once by the logger of the application. The format of the Uvicorn
loggers is also not in JSON format, so remove the handlers of those loggers.

Disable the access logs as logging of the health check pollutes the logs
too much.

Set the logging level of Uvicorn to the level defined in the config.yml.