DEV-2591 Lock usage of MediaHaven client
When the access token is expired, a new one will be automatically requested
making use of the stored refresh token. You can use this refresh token only
once in order to get a new access token and refresh token. Once used, the
refresh token is expired/invalid.

Incoming PREMIS events are handled in an asynchronous way, in a sense that the
actual processing is defined as a "background task". In other words, if the XML
is valid, a response is returned to the client but the actual operation is done
afterwards, in the background, so that the client doesn't need to wait.

These background tasks are apparently handled in a multi-threaded way. It is
possible that the refresh token is being used multiple times simultaneously
when the access token is expired. An easy fix would be to make use of a
lock when using the MediaHaven client.